remove side effects in mutations
